% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFRI_imputation.R
\name{eFRI_imputation}
\alias{eFRI_imputation}
\title{eFRI_imputation}
\usage{
eFRI_imputation(
  segmentation,
  forest_polygon,
  metrics,
  landcover,
  forest_fire,
  forest_harvest,
  ctg,
  lidar_year_field,
  forest_year_field,
  forest_composition_field,
  forest_type_field,
  target_var,
  knn_var
)
}
\arguments{
\item{segmentation}{param}

\item{forest_polygon}{param}

\item{metrics}{param}

\item{landcover}{param}

\item{forest_fire}{param}

\item{forest_harvest}{param}

\item{ctg}{param}

\item{lidar_year_field}{param}

\item{forest_year_field}{param}

\item{forest_composition_field}{param}

\item{forest_type_field}{param}

\item{target_var}{param}

\item{knn_var}{param}
}
\value{
eFRI_imputation
}
\description{
eFRI_imputation
}
\examples{
eFRI_imputation
}
