% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFRI_segmentation.R
\name{eFRI_segmentation}
\alias{eFRI_segmentation}
\title{Segmentation}
\usage{
eFRI_segmentation(
  metrics,
  masks = NULL,
  thresh = 47,
  spec = 0.6,
  spat = 0.6,
  method = "bs",
  clean_nodata = TRUE,
  output_path,
  output_name,
  otb_dir
)
}
\arguments{
\item{metrics}{param}

\item{masks}{param}

\item{thresh}{param}

\item{spec}{param}

\item{spat}{param}

\item{method}{param}

\item{clean_nodata}{param}

\item{output_path}{param}

\item{output_name}{param}

\item{otb_dir}{param}
}
\value{
segmentation
}
\description{
Segmentation
}
\examples{
print("example")
}
