% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forest_age_2019}
\alias{forest_age_2019}
\title{Forest age}
\format{
A SpatRaster of Landsat-derived forest age for Canada 2019 \link[terra:rast]{rast}
}
\source{
Internal
}
\usage{
forest_age_2019
}
\description{
Landsat-derived forest age for Canada 2019
}
\keyword{datasets}
