% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forest_fire_1985_2020}
\alias{forest_fire_1985_2020}
\title{Forest fire}
\format{
A SpatRaster of Landsat-derived forest wildfire disturbances for Canada 1985-2020 \link[terra:rast]{rast}
}
\source{
Internal
}
\usage{
forest_fire_1985_2020
}
\description{
Landsat-derived forest wildfire disturbances for Canada 1985-2020
}
\keyword{datasets}
