% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fri_polygons}
\alias{fri_polygons}
\title{FRI polygons}
\format{
Forest Resources Inventory polygons in \link[sf:sf]{sf}
}
\source{
Internal
}
\usage{
fri_polygons
}
\description{
Forest Resources Inventory polygons
}
\keyword{datasets}
