% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_inputation.R
\name{knn_inputation}
\alias{knn_inputation}
\title{Perform knn imputation}
\usage{
knn_inputation(
  reference_polygons,
  target_polygons,
  knn_variables,
  target_variables,
  k = 5
)
}
\arguments{
\item{reference_polygons}{sf; reference polygons from the \code{sf} package with known values of \code{target_variables}.}

\item{target_polygons}{sf; polygons from the \code{sf} package to impute.}

\item{knn_variables}{character; names of the columns used to find the nearest neighbours.}

\item{target_variables}{character; names of the columns of \code{reference_polygons} to impute.}

\item{k}{numeric; number of nearest neighbours.}
}
\value{
Data frame of the imputed variables with the same length as \code{target_polygons}
}
\description{
Perform knn imputation
}
\examples{
"ex"
}
