% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landcover}
\alias{landcover}
\title{Landcover}
\format{
A SpatRaster of landcover from the Ontario Land Cover Compilation v.2.0 \link[terra:rast]{rast}
}
\source{
Internal
}
\usage{
landcover
}
\description{
Landcover from the Ontario Land Cover Compilation v.2.0
}
\keyword{datasets}
