% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masks}
\alias{masks}
\title{Masks}
\format{
A SpatVector \link[terra:vect]{vect}
}
\source{
Internal
}
\usage{
masks
}
\description{
Masks used to perform segmentation : road and waterbodies
}
\keyword{datasets}
