% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{A subset of metrics used for examples}
\format{
A SpatRaster \link[terra:rast]{rast}
}
\source{
Internal
}
\usage{
metrics
}
\description{
A subset of metrics used for examples : dem, fractional_cover_05_2, relative_topographic_position_100m, rumple_index, sagawi, slope, z_above2, z_cv, z_kurt, z_p95, B6, NDVI, ba_ha, dens, lor, qmdbh, Vmerch_ha
}
\keyword{datasets}
