% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_infos}
\alias{metrics_infos}
\title{Information about all metrics}
\format{
A tibble
}
\source{
Internal
}
\usage{
metrics_infos
}
\description{
A tibble that contains all informations about metrics that can be used : path to data, metric name, description, resolution and type (lidar, sentinel-2 and dendrometric)
}
\keyword{datasets}
