% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_functional_group.R
\name{mutate_functional_group}
\alias{mutate_functional_group}
\title{Extract functional group of stand forest}
\usage{
mutate_functional_group(
  x,
  functional_group_3 = TRUE,
  functional_group_5 = TRUE
)
}
\arguments{
\item{x}{spatial features; polygons of Forest Resources Inventory from the \code{sf} package. species codes with prop, order and forest type must be present before running the function. see \link{mutate_species_prop}, \link{mutate_species_order} and \link{mutate_forest_type}}

\item{functional_group_3}{logical; compute functional group 3 : Softwood, Hardwood and Mixedwood}

\item{functional_group_5}{logical; compute functional group 3 : Jack Pine Dominated, Black Spruce Dominated, Softwood, Hardwood and Mixedwood}
}
\value{
Functionnal group
}
\description{
Extract functional group of stand forest
}
\examples{
library(sf)

fri_polygons_species_prop <- mutate_species_prop(fri_polygons["SPCOMP"], "SPCOMP")

fri_polygons_species_order <- mutate_species_order(fri_polygons_species_prop, "SPCOMP")

fri_polygons_forest_type <- mutate_forest_type(fri_polygons_species_order)

fri_polygons_functional_group <- mutate_functional_group(fri_polygons_forest_type)

fri_polygons_functional_group
}
