% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_perturbation.R
\name{mutate_perturbation}
\alias{mutate_perturbation}
\title{Extract perturbation column from existing perturbation columns with a pattern.}
\usage{
mutate_perturbation(x, col_name, threshold = 80)
}
\arguments{
\item{x}{spatial features; polygons of Forest Resources Inventory from the \code{sf} package. proportion of perturbations columns must be present before running the function. see \link{mutate_proportion}}

\item{col_name}{character; name of the column that contains perturbation proportions}

\item{threshold}{numeric; minimum percentage coverage to consider the perturbation to be significant}
}
\value{
Polygons with the follow perturbation columns : YRDEP, DEPTYPE, DEPPROP
}
\description{
Extract perturbation column from existing perturbation columns with a pattern.
}
\examples{
library(sf)
library(terra)

fri_polygons_proportion <- fri_polygons |>
  mutate_proportion(rast(forest_fire_1985_2020), "forest_fire") |>
  mutate_proportion(rast(forest_harvest_1985_2020), "forest_harvest")

col_name <- grep("^FOREST_FIRE_|^FOREST_HARVEST_", names(fri_polygons_proportion), value = TRUE)

fri_polygons_perturbation <- mutate_perturbation(fri_polygons_proportion, col_name, 80)
}
