% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_proportion.R
\name{mutate_proportion}
\alias{mutate_proportion}
\title{Extract the proportion of each class of a raster in specified polygons}
\usage{
mutate_proportion(
  x,
  y,
  prefix,
  remove_class = c("NaN", "NA", "0"),
  simplify = TRUE,
  keep_all = TRUE
)
}
\arguments{
\item{x}{sf; polygons from the \code{sf} package.}

\item{y}{SpatRaster; categorical raster from the \code{terra} package.}

\item{prefix}{character; prefix of the created columns.}

\item{remove_class}{character; classes to remove, like "NA" or "0".}

\item{simplify}{logical; compute the most frequent class and its proportion.}

\item{keep_all}{logical; keep the proportion columns of each class when \code{simplify = TRUE}.}
}
\value{
Polygons with extracted class proportions as individual columns
}
\description{
Extract the proportion of each class of a raster in specified polygons
}
\examples{
library(sf)
library(terra)

fri_polygons_landcover_prop <- mutate_proportion(fri_polygons, rast(landcover), "landcover")

fri_polygons_landcover_prop
}
