% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{pre_processing}
\alias{pre_processing}
\title{Pre-processing stage}
\usage{
pre_processing(layers, masks = NULL)
}
\arguments{
\item{layers}{SpatRaster. Multiband raster to segment with \link{generic_region_merging}}

\item{masks}{List of SpatVector objects that will be used to mask the raster}
}
\value{
A SpatRaster
}
\description{
Using a multiband raster that will later be used for polygon segmentation in \link{generic_region_merging},
the function masks the unwanted pixel (rivers, lakes, roads, ...), smooth the layers, and normalizes the values to range
in \verb{[0,255]}
}
\examples{
library(terra)
library(purrr)

metrics_pre_processing <- pre_processing(rast(metrics), masks = map(masks, vect))

metrics_pre_processing
}
