% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_variance_explained.R
\name{proportion_variance_explained}
\alias{proportion_variance_explained}
\title{Proportion of the variance of a metric explained by polygons}
\usage{
proportion_variance_explained(x, y)
}
\arguments{
\item{x}{sf; polygons from the \code{sf} package.}

\item{y}{SpatRaster; single metric layer from the \code{terra} package.}
}
\value{
Proportion of explained variance (r2)
}
\description{
Proportion of the variance of a metric explained by polygons
}
\examples{
library(sf)
library(terra)

pve_r2 <- proportion_variance_explained(fri_polygons, rast(metrics)[["z_p95"]])

pve_r2
}
