% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metrics.R
\name{read_metrics}
\alias{read_metrics}
\title{Create a dataframe with all available metrics informations}
\usage{
read_metrics(metric_path)
}
\arguments{
\item{metric_path}{character; folders that contain the metrics as .tif files. Folder names are used as metric type.}
}
\value{
A dataframe with all available metrics informations
}
\description{
Create a dataframe with all available metrics informations
}
\examples{
read_metrics(c("D:/00_Ontario_eFRI/RMF/metrics/lidar",
               "D:/00_Ontario_eFRI/RMF/metrics/sentinel2",
               "D:/00_Ontario_eFRI/RMF/metrics/dendro")) -> metrics_infos
}
