% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subset_area}
\alias{subset_area}
\title{Subset area for testing}
\format{
Subset area in a 2 km radius circle in the centroid of area in \link[sf:sf]{sf}
}
\source{
Internal
}
\usage{
subset_area
}
\description{
Subset area in a 2 km radius circle in the centroid of area for different testing purposes
}
\keyword{datasets}
