% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFRI_attribute_table.R
\name{eFRI_attribute_table}
\alias{eFRI_attribute_table}
\title{eFRI_attribute_table}
\usage{
eFRI_attribute_table(
  segmentation,
  metrics,
  selected_dendrometrics,
  landcover,
  forest_fire,
  forest_harvest,
  forest_age
)
}
\arguments{
\item{segmentation}{param}

\item{metrics}{param}

\item{landcover}{param}

\item{forest_fire}{param}

\item{forest_harvest}{param}

\item{forest_age}{param}
}
\value{
eFRI_attribute_table
}
\description{
eFRI_attribute_table
}
\examples{
eFRI_attribute_table
}
