% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFRI_attribute_table.R
\name{eFRI_attribute_table}
\alias{eFRI_attribute_table}
\title{eFRI_attribute_table}
\usage{
eFRI_attribute_table(
  segmentation,
  metrics,
  summary_metrics,
  landcover,
  forest_fire,
  forest_harvest,
  forest_age
)
}
\arguments{
\item{segmentation}{sf; polygons from the \code{sf} package produced by \code{\link[=eFRI_segmentation]{eFRI_segmentation()}}.}

\item{metrics}{SpatRaster; metrics with one layer per metric from the \code{terra} package.}

\item{summary_metrics}{character; names of the \code{metrics} layers to summarize (median) as columns. \code{NULL} to skip.}

\item{landcover}{SpatRaster; categorical landcover raster from the \code{terra} package.}

\item{forest_fire}{SpatRaster; year of forest fire from the \code{terra} package.}

\item{forest_harvest}{SpatRaster; year of forest harvest from the \code{terra} package.}

\item{forest_age}{SpatRaster; forest age in 2019 from the \code{terra} package.}
}
\value{
Polygons with the eFRI attribute table as columns
}
\description{
eFRI_attribute_table
}
\examples{
eFRI_attribute_table
}
