% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFRI_imputation.R
\name{eFRI_imputation}
\alias{eFRI_imputation}
\title{eFRI_imputation}
\usage{
eFRI_imputation(
  segmentation,
  forest_polygon,
  metrics,
  landcover,
  forest_fire,
  forest_harvest,
  ctg,
  lidar_year_field,
  forest_year_field,
  forest_composition_field,
  forest_type_field,
  target_var,
  knn_var
)
}
\arguments{
\item{segmentation}{sf; polygons from the \code{sf} package produced by \code{\link[=eFRI_segmentation]{eFRI_segmentation()}} or \code{\link[=eFRI_attribute_table]{eFRI_attribute_table()}}.}

\item{forest_polygon}{sf; reference polygons of Forest Resources Inventory from the \code{sf} package.}

\item{metrics}{SpatRaster; metrics with one layer per metric from the \code{terra} package. Must contain \code{z_p95} and \code{z_above2}.}

\item{landcover}{SpatRaster; categorical landcover raster from the \code{terra} package.}

\item{forest_fire}{SpatRaster; year of forest fire from the \code{terra} package.}

\item{forest_harvest}{SpatRaster; year of forest harvest from the \code{terra} package.}

\item{ctg}{sf; LiDAR acquisition tiles from the \code{sf} package.}

\item{lidar_year_field}{character; name of the column of \code{ctg} that contains the LiDAR acquisition year.}

\item{forest_year_field}{character; name of the column of \code{forest_polygon} that contains the inventory year.}

\item{forest_composition_field}{character; name of the column of \code{forest_polygon} that contains species and proportions.}

\item{forest_type_field}{character; name of the column of \code{forest_polygon} that contains the polygon type. Only "FOR" polygons are used.}

\item{target_var}{character; names of the columns of \code{forest_polygon} to impute.}

\item{knn_var}{character; comma-separated metric names used for the knn of each \code{target_var}, of the same length as \code{target_var}.}
}
\value{
\code{segmentation} with imputed variables as individual columns
}
\description{
eFRI_imputation
}
\examples{
eFRI_imputation
}
