% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFRI_segmentation.R
\name{eFRI_segmentation}
\alias{eFRI_segmentation}
\title{Segmentation}
\usage{
eFRI_segmentation(
  metrics,
  masks = NULL,
  thresh = 47,
  spec = 0.6,
  spat = 0.6,
  method = "bs",
  clean_nodata = TRUE,
  output_path,
  output_name,
  otb_dir
)
}
\arguments{
\item{metrics}{SpatRaster; multiband raster to segment from the \code{terra} package.}

\item{masks}{list; SpatVector objects from the \code{terra} package used to mask the raster. \code{NULL} for no mask.}

\item{thresh, spec, spat}{numeric; threshold, spectral weight and spatial weight of GRM. See
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{method}{character; homogeneity criterion of GRM. See
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{clean_nodata}{logical; perform a 3 pixels majority filter to remove small nodata holes.}

\item{output_path}{character; folder where the segmentation raster is saved.}

\item{output_name}{character; file name of the segmentation raster, without extension.}

\item{otb_dir}{character; folder where OTB is installed. Likely "C:/OTB/bin" on Windows.}
}
\value{
Polygons of the segmentation from the \code{sf} package
}
\description{
Segmentation
}
\examples{
print("example")
}
