% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_region_merging.R
\name{generic_region_merging}
\alias{generic_region_merging}
\title{Segmentation of a raster into polygons}
\usage{
generic_region_merging(
  metrics,
  ofile = tempfile(fileext = ".tif"),
  thresh = 50,
  spec = 0.5,
  spat = 0.5,
  method = "bs",
  clean_nodata = TRUE,
  otb_dir
)
}
\arguments{
\item{metrics}{SpatRaster; multiband raster to segment from the \code{terra} package.}

\item{ofile}{character; path where the GRM raster is saved.}

\item{thresh, spec, spat}{numeric; threshold, spectral weight and spatial weight of GRM. See
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{method}{character; homogeneity criterion of GRM. See
\url{https://www.orfeo-toolbox.org/CookBook-8.0/Applications/app_GenericRegionMerging.html}}

\item{clean_nodata}{logical; perform a 3 pixels majority filter to remove small nodata holes.}

\item{otb_dir}{character; folder where OTB is installed. Likely "C:/OTB/bin" on Windows.}
}
\value{
Polygons of the segmentation from the \code{sf} package
}
\description{
\code{generic_region_merging} calls the GenericRegionMerging software from Orfeo ToolBox (OTB).
OTB must be installed on your computer first. See \url{https://www.orfeo-toolbox.org/CookBook-7.0/index.html}
}
\examples{
\dontrun{
library(terra)
library(sf)

}
}
