% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_inputation.R
\name{knn_inputation}
\alias{knn_inputation}
\title{Perform knn imputation}
\usage{
knn_inputation(
  reference_polygons,
  target_polygons,
  knn_variables,
  target_variables,
  k = 5
)
}
\arguments{
\item{reference_polygons}{spatial features; reference polygons from the \code{sf} package.}

\item{target_polygons}{spatial features; target polygons from the \code{sf} package.}

\item{knn_variables}{Variables to use}

\item{target_variables}{Variables to infer}

\item{k}{The maximum number of nearest neighbours to compute}
}
\value{
table of the imputed variable of the same lenght as the target polygons
}
\description{
Perform knn imputation
}
\examples{
"ex"
}
