% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_age.R
\name{mutate_age}
\alias{mutate_age}
\title{Extract age of forest stand}
\usage{
mutate_age(x, y, current_year, target_year, fun)
}
\arguments{
\item{x}{sf; polygons from the \code{sf} package.}

\item{y}{SpatRaster; forest age from the \code{terra} package.}

\item{current_year}{numeric; year of the forest age raster \code{y}.}

\item{target_year}{numeric; year for which the age is computed, usually the current year.}

\item{fun}{character; summary function of \code{exactextractr::exact_extract()} (e.g. "mean", "median").}
}
\value{
Polygons with extracted age as a single column
}
\description{
Extract age of forest stand
}
\examples{
library(sf)
library(terra)

fri_polygons_age <- mutate_age(fri_polygons, rast(forest_age_2019), 2019, 2025, "mean")

fri_polygons_age
}
