% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_age.R
\name{mutate_age}
\alias{mutate_age}
\title{Extract age of forest stand}
\usage{
mutate_age(x, y, current_year, target_year, fun)
}
\arguments{
\item{x}{spatial features; polygons from the \code{sf} package of forest stands.}

\item{y}{forest age; SpatRaster from the \code{terra} package.}

\item{current_year}{year of the SpatRaster forest age.}

\item{target_year}{target year desired; usually the current year.}

\item{fun}{function to summarize the extracted age by polygon.}
}
\value{
#' Polygons with extracted age as a single column
}
\description{
Extract age of forest stand
}
\examples{
library(sf)
library(terra)

fri_polygons_age <- mutate_age(fri_polygons, rast(forest_age_2019), 2019, 2025, "mean")

fri_polygons_age
}
