% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_centroid.R
\name{mutate_centroid}
\alias{mutate_centroid}
\title{Extract the X and Y coordinates of the polygon centroids}
\usage{
mutate_centroid(x)
}
\arguments{
\item{x}{spatial features; polygons from the \code{sf} package.}
}
\value{
Polygons with extracted X and Y coordinates as individuals columns
}
\description{
Extract the X and Y coordinates of the polygon centroids
}
\examples{
library(sf)

fri_polygons_geometry <- st_geometry(fri_polygons)

fri_polygons_centroids <- mutate_centroid(fri_polygons_geometry)

fri_polygons_centroids
}
