% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_disturbance.R
\name{mutate_disturbance}
\alias{mutate_disturbance}
\title{Extract disturbance column from existing disturbance columns with a pattern.}
\usage{
mutate_disturbance(x, col_name, threshold = 80)
}
\arguments{
\item{x}{spatial features; polygons of Forest Resources Inventory from the \code{sf} package. proportion of disturbances columns must be present before running the function. see \link{mutate_proportion}}

\item{col_name}{character; name of the column that contains disturbance proportions}

\item{threshold}{numeric; minimum percentage coverage to consider the disturbance to be significant}
}
\value{
Polygons with the follow disturbance columns : YRDEP, DEPTYPE, DEPPROP
}
\description{
Extract disturbance column from existing disturbance columns with a pattern.
}
\examples{
library(sf)
library(terra)

fri_polygons_proportion <- fri_polygons |>
  mutate_proportion(rast(forest_fire_1985_2020), "forest_fire") |>
  mutate_proportion(rast(forest_harvest_1985_2020), "forest_harvest")

col_name <- grep("^FOREST_FIRE_|^FOREST_HARVEST_", names(fri_polygons_proportion), value = TRUE)

fri_polygons_disturbance <- mutate_disturbance(fri_polygons_proportion, col_name, 80)
}
