% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_forest_type.R
\name{mutate_forest_type}
\alias{mutate_forest_type}
\title{Extract forest type proportions from species proportion columns}
\usage{
mutate_forest_type(x)
}
\arguments{
\item{x}{sf; polygons from the \code{sf} package. Species proportion columns must be present, see \code{\link[=mutate_species_prop]{mutate_species_prop()}}.}
}
\value{
Polygons with extracted forest type proportions as two individual columns: PROP_CONIFEROUS and PROP_DECIDUOUS
}
\description{
Extract forest type proportions from species proportion columns
}
\examples{
library(sf)

fri_polygons_species_prop <- mutate_species_prop(fri_polygons["SPCOMP"], "SPCOMP")

fri_polygons_forest_type <- mutate_forest_type(fri_polygons_species_prop)

fri_polygons_forest_type
}
