% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_functional_group.R
\name{mutate_functional_group}
\alias{mutate_functional_group}
\title{Extract functional group of stand forest}
\usage{
mutate_functional_group(
  x,
  functional_group_3 = TRUE,
  functional_group_5 = TRUE
)
}
\arguments{
\item{x}{sf; polygons from the \code{sf} package. Species proportion, species order and forest type columns must be present, see \code{\link[=mutate_species_prop]{mutate_species_prop()}}, \code{\link[=mutate_species_order]{mutate_species_order()}} and \code{\link[=mutate_forest_type]{mutate_forest_type()}}.}

\item{functional_group_3}{logical; compute the 3 functional groups: Softwood, Hardwood and Mixedwood.}

\item{functional_group_5}{logical; compute the 5 functional groups: Jack Pine Dominated, Black Spruce Dominated, Mixed Conifer, Hardwood and Mixedwood.}
}
\value{
Polygons with functional groups as individual columns: FUNCTIONAL_GROUP_3 and FUNCTIONAL_GROUP_5
}
\description{
Extract functional group of stand forest
}
\examples{
library(sf)

fri_polygons_species_prop <- mutate_species_prop(fri_polygons["SPCOMP"], "SPCOMP")

fri_polygons_species_order <- mutate_species_order(fri_polygons_species_prop, "SPCOMP")

fri_polygons_forest_type <- mutate_forest_type(fri_polygons_species_order)

fri_polygons_functional_group <- mutate_functional_group(fri_polygons_forest_type)

fri_polygons_functional_group
}
