% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_metrics.R
\name{mutate_metrics}
\alias{mutate_metrics}
\title{Extract one or multiple metrics from polygons}
\usage{
mutate_metrics(x, y, fun = "median")
}
\arguments{
\item{x}{spatial features; polygons from the \code{sf} package.}

\item{y}{metrics; SpatRaster with one or more layers from the \code{terra} package.}

\item{fun}{function to summarize the extracted metrics by polygon.}
}
\value{
Polygons with extracted metrics as individuals columns
}
\description{
Extract one or multiple metrics from polygons
}
\examples{
library(sf)
library(terra)

fri_polygons_metrics <- mutate_metrics(fri_polygons, rast(metrics))

fri_polygons_metrics
}
