% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_perturbation.R
\name{mutate_perturbation}
\alias{mutate_perturbation}
\title{Extract perturbation column from existing perturbation coloumns with a pettern.}
\usage{
mutate_perturbation(x, col_name, threshold = 80)
}
\arguments{
\item{x}{spatial features; polygons of Forest Resources Inventory from the \code{sf} package. proportion of perturbations columns must be present before running the function. see \link{mutate_proportion}}

\item{col_name}{character; name of the column that contains perturbation proportions}

\item{threshold}{numeric; minimum percentage coverage to consider the perturbation to be significant}
}
\value{
Polygons with the follow perturbation columns : YRDEP, DEPTYPE, DEPPROP
}
\description{
Extract perturbation column from existing perturbation coloumns with a pettern.
}
\examples{
x <- fri_polygons \%>\%
  mutate_proportion(rast(forest_fire_1985_2020), "forest_fire") \%>\%
  mutate_proportion(rast(forest_harvest_1985_2020), "forest_harvest")

col_name <- grep("^FOREST_FIRE_|^FOREST_HARVEST_", names(x), value = TRUE)

threshold <- 20
}
