% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_prop_forested.R
\name{mutate_prop_forested}
\alias{mutate_prop_forested}
\title{Extract the proportion of the polygon that are forested}
\usage{
mutate_prop_forested(
  x,
  y,
  forested_class = c("Sparse_Treed", "Treed_Upland", "Deciduous_Treed", "Mixed_Treed",
    "Coniferous_Treed", "Plantations_Treed_Cultivated", "Hedge_Rows",
    "Tallgrass_Woodland")
)
}
\arguments{
\item{x}{spatial features; polygons of Forest Resources Inventory from the \code{sf} package.}

\item{y}{raster; Landcover SpatRaster from the \code{terra} package.}

\item{forested_class}{vector; all class of the landcover SpatRaster that are forested}
}
\value{
Column "PROPFORESTED"
}
\description{
Extract the proportion of the polygon that are forested
}
\examples{
library(sf)
library(terra)

fri_polygons_prop_forested <- mutate_prop_forested(fri_polygons, rast(landcover), forested_class = c("Sparse_Treed",
                                                                                                     "Treed_Upland",
                                                                                                     "Deciduous_Treed",
                                                                                                     "Mixed_Treed",
                                                                                                     "Coniferous_Treed",
                                                                                                     "Plantations_Treed_Cultivated",
                                                                                                     "Hedge_Rows",
                                                                                                     "Tallgrass_Woodland"))

fri_polygons_prop_forested
}
