% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_proportion.R
\name{mutate_proportion}
\alias{mutate_proportion}
\title{Extract the proportion of each class of a raster in specified polygons}
\usage{
mutate_proportion(x, y, prefix, simplify = TRUE, keep_all = TRUE)
}
\arguments{
\item{x}{spatial features; polygons of Forest Resources Inventory from the \code{sf} package.}

\item{y}{raster; SpatRaster with categorical values from the \code{terra} package.}

\item{prefix}{character; prefix use for created column for each class}

\item{simplify}{logical; compute the most frequent class and its proportion}

\item{keep_all}{logical; keep all column of individual class}
}
\value{
Polygons with extracted class proportions as individuals columns
}
\description{
Extract the proportion of each class of a raster in specified polygons
}
\examples{
library(sf)
library(terra)

fri_polygons_landcover_prop <- mutate_proportion(fri_polygons, rast(landcover), "landcover")

fri_polygons_landcover_prop
}
