% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_species_order.R
\name{mutate_species_order}
\alias{mutate_species_order}
\title{Separate a column of species composition into species orders}
\usage{
mutate_species_order(x, col_name)
}
\arguments{
\item{x}{sf; polygons from the \code{sf} package.}

\item{col_name}{character; name of the column that contains species and proportions.}
}
\value{
Polygons with extracted species orders as individual columns
}
\description{
Separate a column of species composition into species orders
}
\examples{
library(sf)

fri_polygons_species_order <- mutate_species_order(fri_polygons["SPCOMP"], "SPCOMP")

fri_polygons_species_order
}
