% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_species_prop.R
\name{mutate_species_prop}
\alias{mutate_species_prop}
\title{Separate a column of species composition into proportions}
\usage{
mutate_species_prop(x, col_name, update_names = TRUE)
}
\arguments{
\item{x}{sf; polygons from the \code{sf} package.}

\item{col_name}{character; name of the column that contains species and proportions.}

\item{update_names}{logical; replace species proportion columns that already exist.}
}
\value{
Polygons with extracted species proportions as individual columns
}
\description{
Separate a column of species composition into proportions
}
\examples{
library(sf)

fri_polygons_species_prop <- mutate_species_prop(fri_polygons["SPCOMP"], "SPCOMP")

fri_polygons_species_prop
}
