% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_species_prop.R
\name{mutate_species_prop}
\alias{mutate_species_prop}
\title{Separate a column of species composition into multiple columns}
\usage{
mutate_species_prop(x, col_name)
}
\arguments{
\item{x}{spatial features; polygons of Forest Resources Inventory from the \code{sf} package.}

\item{col_name}{character; name of the column that contains species and proportions}
}
\value{
Polygons with extracted specie proportions as individuals columns
}
\description{
Separate a column of species composition into multiple columns
}
\examples{
library(sf)

fri_polygons_species <- mutate_species_prop(fri_polygons["SPCOMP"], "SPCOMP")

fri_polygons_species
}
