% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{pre_processing}
\alias{pre_processing}
\title{Pre-processing stage}
\usage{
pre_processing(metrics, masks = NULL)
}
\arguments{
\item{metrics}{SpatRaster; multiband raster to segment with \code{\link[=generic_region_merging]{generic_region_merging()}} from the \code{terra} package.}

\item{masks}{list; SpatVector objects from the \code{terra} package used to mask the raster. \code{NULL} for no mask.}
}
\value{
Pre-processed SpatRaster from the \code{terra} package
}
\description{
Using a multiband raster that will later be used for polygon segmentation in \link{generic_region_merging},
the function masks the unwanted pixel (rivers, lakes, roads, ...), smooth the layers, and normalizes the values to range
in \verb{[0,255]}
}
\examples{
library(terra)

metrics_pre_processing <- pre_processing(rast(metrics), masks = lapply(masks, vect))

metrics_pre_processing
}
