% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_variance_explained.R
\name{proportion_variance_explained}
\alias{proportion_variance_explained}
\title{Title}
\usage{
proportion_variance_explained(x, y)
}
\arguments{
\item{x}{spatial features; polygons from the \code{sf} package.}

\item{y}{metric; SpatRaster with one from the \code{terra} package.}
}
\value{
proportion of explained variance in r2
}
\description{
Title
}
\examples{
library(sf)
library(terra)

pve_r2 <- proportion_variance_explained(fri_polygons, rast(metrics)[["z_p95"]])

pve_r2
}
